requirements.txt
SQLAlchemy==2.0.24
aiogram
python-dotenv